Move all Models into Django Project
